Restructure package as derivkit with cleaned layout and notebooks
