+++ .gitignore
@@ -2,3 +2,4 @@
 *__pycache__
 logs/
 plots_tests/
+notebooks/logs/
