+++ derivkit/__init__.py
@@ -1,0 +1,4 @@
+from .hybrid import HybridDerivativeCalculator
+from .plotter import DerivativePlotter
+
+__all__ = ["HybridDerivativeCalculator", "DerivativePlotter"]
