+++ pyproject.toml
@@ -1,0 +1,35 @@
+[build-system]
+requires = ["setuptools", "wheel"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "derivkit"
+version = "0.1.0"
+description = "Toolkit for estimating derivatives using hybrid polynomial fitting and stencil methods"
+readme = "README.md"
+license = { text = "MIT" }
+requires-python = ">=3.8"
+
+authors = [
+  { name = "Nikolina Šarčević" },
+  { name = "Matthijs van der Wild" }
+]
+
+dependencies = [
+  "numpy",
+  "matplotlib",
+  "seaborn",
+  "scipy",
+  "numdifftools"
+]
+
+keywords = ["derivatives", "numerical differentiation", "STEM method", "Fisher forecasting"]
+
+classifiers = [
+  "Programming Language :: Python :: 3",
+  "License :: OSI Approved :: MIT License",
+  "Operating System :: OS Independent"
+]
+
+[tool.setuptools.packages.find]
+where = ["."]
